// src/interfaces/proyectosInterface.tsx
// Generated by https://quicktype.io

export interface InterfacesProyectos {
    // map(arg0: (projec: InterfacesProyectos) => JSX.Element): import("react").ReactNode;
    image:       Image;
    _id:         string;
    title:       string;
    descripcion: string;
    tech:        string;
    createdAt:   string;
    updatedAt:   string;
    github?:     string;
    proyectUrl?: string;
    
    
    
}

export interface Image {
    url:       string;
    public_id: string;
}